Move SuddenDeath colors to styles + edit styles data
