+++ TODO.txt
@@ -1,7 +1,6 @@
 TODO
 ---------------------------------------------------------------------------------------
 >>	implementovat Sudden death mode
-	- presunout barvy jako promenne do color.ts
 	- primarne vyjmout ty akce, kde by postava zustala na highlighted poli
 	- pri Evasion prioritizovat pole, ktere nejsou highlighted
